+++ .gitignore
@@ -8,3 +8,4 @@ buildtarget
 .idea
 output
 out
+*.log
